Put the menu options in a separate class
