+++ OpenKh.Game/Menu/MenuTemplate.cs
@@ -1,0 +1,24 @@
+﻿namespace OpenKh.Game.Menu
+{
+    public class MenuTemplate : IMenu
+    {
+        public int SelectedOption { get; private set; }
+
+        public void Open()
+        {
+
+        }
+
+        public void Close()
+        {
+        }
+
+        public void Update(double deltaTime)
+        {
+        }
+
+        public void Draw()
+        {
+        }
+    }
+}
